requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyascii==0.3.2
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
bangla==0.0.2
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
coqpit==0.0.17
cycler==0.11.0
Cython==0.29.28
dateparser==1.1.8
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.11.0
Flask==2.3.1
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
g2pkk==0.1.2
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gTTS==2.3.1
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
inflect==5.6.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonlines==1.2.0
jupyter-client==8.2.0
jupyter-core==5.3.0
kiwisolver==1.4.4
lazy-loader==0.2
librosa==0.10.0.post2
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mecab-python3==1.0.5
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.56.4
numpy==1.21.6
openai==0.27.4
openai-whisper==20230314
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pooch==1.6.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycodestyle==2.10.0
pycparser==2.21
pydub==0.25.1
Pygments==2.15.1
pynndescent==0.5.10
pyparsing==3.0.9
pypinyin==0.48.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
simpleaudio==1.0.4
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
SpeechRecognition==3.10.0
stack-data==0.6.2
sympy==1.11.1
tensorboardX==2.6
threadpoolctl==3.1.0
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
tornado==6.3
tqdm==4.65.0
trainer==0.0.20
traitlets==5.9.0
TTS==0.13.3
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
umap-learn==0.5.1
unidic-lite==1.0.8
urllib3==1.26.15
wavio==0.0.7
wcwidth==0.2.6
Werkzeug==2.3.1
yarl==1.8.2
zipp==3.15.0
